+++ src/Serializer/LineItem/LineItemContainerSerializer.php
@@ -27,17 +27,26 @@ use OAT\Library\Lti1p3Ags\Factory\LineItem\LineItemFactoryInterface;
 use OAT\Library\Lti1p3Ags\Model\LineItem\LineItemCollection;
 use OAT\Library\Lti1p3Ags\Model\LineItem\LineItemContainer;
 use OAT\Library\Lti1p3Ags\Model\LineItem\LineItemContainerInterface;
+use OAT\Library\Lti1p3Ags\Serializer\JsonSerializer;
+use OAT\Library\Lti1p3Ags\Serializer\JsonSerializerInterface;
 use OAT\Library\Lti1p3Core\Exception\LtiException;
 use OAT\Library\Lti1p3Core\Exception\LtiExceptionInterface;
+use RuntimeException;
 
 class LineItemContainerSerializer implements LineItemContainerSerializerInterface
 {
     /** @var LineItemFactoryInterface */
     private $lineItemFactory;
 
-    public function __construct(?LineItemFactoryInterface $factory = null)
-    {
+    /** @var JsonSerializerInterface */
+    private $jsonSerializer;
+
+    public function __construct(
+        ?LineItemFactoryInterface $factory = null,
+        ?JsonSerializerInterface $jsonSerializer = null
+    ) {
         $this->lineItemFactory = $factory ?? new LineItemFactory();
+        $this->jsonSerializer = $jsonSerializer ?? new JsonSerializer();
     }
 
     /**
@@ -45,15 +54,13 @@ class LineItemContainerSerializer implements LineItemContainerSerializerInterfac
      */
     public function serialize(LineItemContainerInterface $container): string
     {
-        $json = json_encode($container);
-
-        if (JSON_ERROR_NONE !== json_last_error()) {
+        try {
+            return $this->jsonSerializer->serialize($container);
+        } catch (RuntimeException $exception) {
             throw new LtiException(
-                sprintf('Error during line item container serialization: %s', json_last_error_msg())
+                sprintf('Error during line item container serialization: %s', $exception->getMessage())
             );
         }
-
-        return $json;
     }
 
     /**
@@ -61,23 +68,23 @@ class LineItemContainerSerializer implements LineItemContainerSerializerInterfac
      */
     public function deserialize(string $data): LineItemContainerInterface
     {
-        $data = json_decode($data, true);
-
-        if (JSON_ERROR_NONE !== json_last_error()) {
+        try {
+            $deserializedData = $this->jsonSerializer->deserialize($data);
+        } catch (RuntimeException $exception) {
             throw new LtiException(
-                sprintf('Error during line item container deserialization: %s', json_last_error_msg())
+                sprintf('Error during line item container deserialization: %s', $exception->getMessage())
             );
         }
 
         $collection = new LineItemCollection();
 
-        foreach ($data['lineItems'] ?? [] as $lineItemData) {
+        foreach ($deserializedData['lineItems'] ?? [] as $lineItemData) {
             $collection->add($this->lineItemFactory->create($lineItemData));
         }
 
         return new LineItemContainer(
             $collection,
-            $data['relationLink'] ?? null
+            $deserializedData['relationLink'] ?? null
         );
     }
 }