+++ tests/Unit/Serializer/Result/ResultSerializerTest.php
@@ -24,26 +24,21 @@ namespace OAT\Library\Lti1p3Ags\Tests\Unit\Serializer\Result;
 
 use Carbon\Carbon;
 use OAT\Library\Lti1p3Ags\Model\Result\ResultInterface;
-use OAT\Library\Lti1p3Ags\Model\Score\ScoreInterface;
+use OAT\Library\Lti1p3Ags\Serializer\JsonSerializerInterface;
 use OAT\Library\Lti1p3Ags\Serializer\Result\ResultSerializer;
-use OAT\Library\Lti1p3Ags\Serializer\Result\ResultSerializerInterface;
 use OAT\Library\Lti1p3Ags\Tests\Traits\AgsDomainTestingTrait;
 use OAT\Library\Lti1p3Core\Exception\LtiExceptionInterface;
 use PHPUnit\Framework\TestCase;
+use RuntimeException;
 
 class ResultSerializerTest extends TestCase
 {
     use AgsDomainTestingTrait;
 
-    /** @var ResultSerializerInterface */
-    private $subject;
-
     protected function setUp(): void
     {
         $now = Carbon::now()->setMicro(0);
         Carbon::setTestNow($now);
-
-        $this->subject = new ResultSerializer();
     }
 
     protected function tearDown(): void
@@ -53,15 +48,19 @@ class ResultSerializerTest extends TestCase
 
     public function testSerializeForFailure(): void
     {
-        $invalidContainer = $this->createMock(ResultInterface::class);
-        $invalidContainer->expects($this->once())
-            ->method('jsonSerialize')
-            ->willReturn(NAN); // Note: NaN cannot be JSON encoded
+        $resultMock = $this->createMock(ResultInterface::class);
+        $serializerMock = $this->createMock(JsonSerializerInterface::class);
+        $subject = new ResultSerializer(null, $serializerMock);
+
+        $serializerMock->expects($this->once())
+            ->method('serialize')
+            ->with($resultMock)
+            ->willThrowException(new RuntimeException('some error'));
 
         $this->expectException(LtiExceptionInterface::class);
         $this->expectExceptionMessage('Error during result serialization');
 
-        $this->subject->serialize($invalidContainer);
+        $subject->serialize($resultMock);
     }
 
     public function testSerializeForSuccess(): void
@@ -70,7 +69,7 @@ class ResultSerializerTest extends TestCase
 
         $this->assertEquals(
             json_encode($result->jsonSerialize()),
-            $this->subject->serialize($result)
+            (new ResultSerializer())->serialize($result)
         );
     }
 
@@ -80,15 +79,23 @@ class ResultSerializerTest extends TestCase
 
         $this->assertEquals(
             $result,
-            $this->subject->deserialize(json_encode($result->jsonSerialize()))
+            (new ResultSerializer())->deserialize(json_encode($result->jsonSerialize()))
         );
     }
 
     public function testDeserializeFailure(): void
     {
+        $serializerMock = $this->createMock(JsonSerializerInterface::class);
+        $subject = new ResultSerializer(null, $serializerMock);
+
+        $serializerMock->expects($this->once())
+            ->method('deserialize')
+            ->with('{')
+            ->willThrowException(new RuntimeException('some error'));
+
         $this->expectException(LtiExceptionInterface::class);
         $this->expectExceptionMessage('Error during result deserialization');
 
-        $this->subject->deserialize('{');
+        $subject->deserialize('{');
     }
 }